Allow setting multiple systems to lock
This commit allows us to specify multiple systems that we'd want to
lock.

This allows _any_ system to be locked by simply using a new redis key
for each system. The drawback is that if someone has a typo on the
system name we'll simply "lock" that one and the person won't be placed
in the queue that they wanted.