Adding `service` model
This commit adds a new model that is in charge of the logic to
lock/unlock services and handle the queue of users waiting to acquire
the lock.

This works by using two sorted sets in Redis:

* The service set holds the users that have requested to acquire the
lock, sorted by request time. The first user in this set is the current
lock holder.

Since users can request to lock a service for a custom amount of time,
the values of this set also include that time in seconds:

Running the following Ruby code

```ruby
s = Service.new("api")
s.lock(user: "nhocki", seconds: 30)
s.lock(user: "sergio", seconds: 60)
```

results in having the following values in Redis:

```
127.0.0.1:6379> ZRANGE services:api:queue 0 -1
1) "nhocki:30"
2) "sergio:60"
```

The second set is `services:__locked__` and holds the name of **every
service that is currently locked**. The score for each service name is
the **timestamp when the service should be unlocked**.

```
127.0.0.1:6379> ZRANGE services:__locked__ 0 -1 WITHSCORES
1) "api"
2) "1555792112"
```

This allows us to automatically unlock services by expiring locks of
services whose score is lower than the current timestamp.